fix: message rename and remove defineprops and defineemits import
