+++ apps/client/components/main/Message/message.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup name="Message">
 import { onMounted, ref } from "vue";
-import { Type } from "./Message";
+import { Type } from "./useMessage";
 
 const { type = Type.SUCCESS, duration = 2000 } = defineProps<{
   type?: Type;
@@ -112,3 +112,4 @@ onMounted(() => {
   transform: translate(-50%, 0);
 }
 </style>
+./useMessage
