+++ apps/client/pages/Auth/FormInput.vue
@@ -23,8 +23,6 @@
 </template>
 
 <script setup lang="ts">
-import { defineEmits, defineProps } from "vue";
-
 const props = defineProps({
   label: String,
   name: String,
