+++ apps/client/pages/Auth/Login.vue
@@ -83,3 +83,4 @@ const handleLogin = handleSubmit(async (values) => {
   }
 });
 </script>
+~/components/main/Message/useMessage
