+++ apps/client/pages/Auth/SignUp.vue
@@ -112,3 +112,4 @@ const handleRegister = handleSubmit(async (values) => {
   }
 });
 </script>
+~/components/main/Message/useMessage
